requirements.txt
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.5
urllib3==2.5.0
python-dotenv==1.1.0
